requirements.txt
asgiref
autopep8
django==2.2.10
pycodestyle
pytz
sqlparse
Unipath~=1.1
dj-database-url
python-decouple
gunicorn
whitenoise
paho-mqtt~=1.5.1
